feat(byos): complete modular refactor and restore functionality
