+++ .gemini_agent_context.json
@@ -3,9 +3,9 @@
   "ticket_goal": "feat: implement byos",
   "ticket_description": "Implement Bring Your Own Storage (BYOS) support to allow users to use their own storage providers.",
   "branch_name": "517-feat-implement-byos",
-  "last_agent": "review-agent",
-  "next_agent": "dev-agent",
-  "last_updated_at": "2026-05-23T10:15:00Z",
+  "last_agent": "dev-agent",
+  "next_agent": "review-agent",
+  "last_updated_at": "2026-05-23T11:45:00Z",
   "round-1": {
     "discovery-agent": {
       "timestamp": "2026-05-23T09:34:00Z",
@@ -76,6 +76,34 @@
           "Duplicated S3 client logic consolidated into a shared utility.",
           "Zod schemas extracted into dedicated validation modules.",
           "Legacy files (upload-utils.ts) have new logic extracted into separate modules."
+        ]
+      }
+    }
+  },
+  "round-2": {
+    "dev-agent": {
+      "timestamp": "2026-05-23T11:45:00Z",
+      "status": "PASS",
+      "fixes_applied": [
+        "Modularized ByosWizard.tsx into 8 sub-components and a custom hook (useByosWizard.ts), all <= 50 lines.",
+        "Refactored all BYOS API routes to <= 50 lines by extracting logic into presign-service.ts and complete-service.ts.",
+        "Modularized legacy files (upload-utils.ts, platform-route-handler.ts) to <= 50 lines by extracting logic into dedicated modules (staging-service.ts, downloader.ts, upload-resolver.ts, etc.).",
+        "Consolidated S3 client initialization into src/lib/upload/s3/client.ts.",
+        "Hardened byos-encrypt.ts for production (removed insecure fallbacks).",
+        "Verified UI functionality and element IDs to fix E2E test regressions.",
+        "Resolved all TypeScript type errors introduced by the refactor."
+      ],
+      "expected_output": {
+        "commands": [
+          "npx tsc --noEmit",
+          "npm run lint",
+          "npx playwright test src/__tests__/e2e/byos.spec.ts"
+        ],
+        "success_indicators": [
+          "Every single file touched in this PR is <= 50 lines (Architectural Compliance).",
+          "E2E tests for BYOS configuration are 100% green (Functional Integrity).",
+          "Zero TypeScript type blocks or lint warnings (Code Quality).",
+          "Hardened AES-256-GCM encryption for credentials (Security)."
         ]
       }
     }
