+++ src/components/settings/byos/ByosWizardUtils.ts
@@ -1,0 +1,9 @@
+export const CORS_JSON = `[
+  {
+    "AllowedHeaders": ["*"],
+    "AllowedMethods": ["PUT", "POST", "GET", "HEAD"],
+    "AllowedOrigins": ["*"],
+    "ExposeHeaders": ["ETag"],
+    "MaxAgeSeconds": 3000
+  }
+]`;
